# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: 公论 (GongLun) API
  description: 公论结构化讨论系统后端 API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /topics:
    get:
      tags: [话题]
      summary: 获取话题列表
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
            enum: [createdAt, updatedAt]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 话题列表

    post:
      tags: [话题]
      summary: 创建话题
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  maxLength: 80
                body:
                  type: string
                  maxLength: 5000
      responses:
        '201':
          description: 话题创建成功
        '401':
          description: 未认证
  /topics/{topicId}:
    get:
      tags: [话题]
      summary: 获取单个话题
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 话题详情
        '404':
          description: 话题不存在

    patch:
      tags: [话题]
      summary: 更新话题状态
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: topicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [OPEN, ARCHIVED]
      responses:
        '200':
          description: 更新成功
        '403':
          description: 无权限
        '404':
          description: 话题不存在
...